requirements.txt
discord.py
requests
aiohttp
Pillow
PyNaCl
